requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
boto3==1.34.59
botocore==1.34.59
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
distro==1.9.0
Django==4.2.9
django-cors-headers==4.3.1
django-rest-swagger==2.2.0
django-storages==1.14.2
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.2.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
JPype1==1.5.0
konlpy==0.6.0
lxml==5.1.0
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
neo4j==5.18.0
numpy==1.26.4
openai==1.30.5
openapi-codec==1.3.2
packaging==23.2
pillow==10.2.0
pydantic==2.7.2
pydantic_core==2.18.3
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.0
scikit-learn==1.4.1.post1
scipy==1.12.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
threadpoolctl==3.3.0
tqdm==4.66.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.18
